# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Blog App Server",
    "description": "This is a swagger documentation for Blog App based on the OpenAPI 3.0 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io).",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  },
  "servers": [
    {
      "url": "http://localhost:3222/"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "Operations with the users"
    },
    {
      "name": "Posts",
      "description": "Operations with the posts"
    },
    {
      "name": "Files",
      "description": "Operations with the files"
    }
  ],
  "paths": {
    "/auth/me/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get information about user",
        "operationId": "getUserInfo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "user Info by ID",
            "example": "42abcd2b-8b9c-4af9-88f7-0bc180cf74b4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register user",
        "operationId": "geristerUser",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Login User",
        "operationId": "loginUser",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Login"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/posts": {
      "get": {
        "summary": "Get all posts",
        "operationId": "getPosts",
        "tags": [
          "Posts"
        ],
        "responses": {
          "200": {
            "description": "Get posts",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add post",
        "operationId": "createPost",
        "tags": [
          "Posts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully added",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Incorect input data",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "summary": "Get post by id",
        "operationId": "getPostId",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "indificator of post",
            "example": "42abcd2b-8b9c-4af9-88f7-0bc180cf74b4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update post by id",
        "operationId": "updatePost",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Indificator of post",
            "example": "42abcd2b-8b9c-4af9-88f7-0bc180cf74b4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success update",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Incorrect input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete post by id",
        "operationId": "deletePost",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Indificator of post",
            "example": "42abcd2b-8b9c-4af9-88f7-0bc180cf74b4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success update",
            "content": {}
          }
        }
      }
    },
    "/upload": {
      "post": {
        "operationId": "uploadFile",
        "tags": [
          "Files"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success update",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "required": [
          "id",
          "fullName",
          "email"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dslkjsl-12n12-1wskls3"
          },
          "fullName": {
            "type": "string",
            "example": "Max"
          },
          "email": {
            "type": "string",
            "example": "max@gmail.com"
          },
          "photoUrls": {
            "type": "string"
          }
        },
        "items": {
          "$ref": "#/components/schemas/User"
        }
      },
      "Login": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "max@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "Max123"
          }
        },
        "items": {
          "$ref": "#/components/schemas/Login"
        }
      },
      "Post": {
        "required": [
          "title",
          "text"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dslkjsl-12n12-1wskls3"
          },
          "title": {
            "type": "string",
            "example": "Hakuna Matata!"
          },
          "tags": {
            "type": "array",
            "example": ["Smile","Work"]
          },
          "text": {
            "type": "string",
            "example": "What a wonderful phrase..."
          },
          "imageUrl": {
            "type": "string"
          }
        },
        "items": {
          "$ref": "#/components/schemas/Post"
        }
      },
      "File": {
        "items": {
          "$ref": "#/components/schemas/File"
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}